backend/requirements.txt
 # Dependencies
crewai
langchain-google-genai
google-generativeai
flask
psycopg2-binary
sentence-transformers
python-dotenv
PyMuPDF